Research Data/Experimental/Experimental-2954648d4d35e4014a65fee1737259e7e5ef525c/nlp/tools/gemini-api/requirements.txt
streamlit
openai
pymongo
voyageai
python-dotenv
langchain-text-splitters